Cleaned up app.js with a switch
Added a switch statement to get rid of a few conditionals.  Unfortunately we have a lot of conditionals inside each conditionals.  So it only got rid of about 10 lines of code.  

Edited a few comparison statements to have three '=' signs instead of 2 in the appropriate spots.  

In cooktime and preptime got rid of the unnecessary variable and made it a return statement.  

Also, deleted the config file that must live in develop as it was in this branch after branching from develop on the most recent version.