Deleted the config file from repo and minor adjustments
Deleted the config.txt file as it should not be in the repo.  

Changed a lot of the comparisons to have the 3 '=' signs per javascripts request to not cause unexpected type coercion.  

changed the previous step in app.js to an else if from an if statement.

Changed the return statement in cook time.