Cleaning up app.js
Removed the functionality from app.js and move it all into the Intent specific functions so that the switch statement is just a router to retrieve the response message.

Removed all the duplicate code in all the step functions to use the mongo helper function.  

Next is to remove the functionality from the steps in the app.js switch statement.  They are dependent on index variables, so need to figure out how that is working to add it all into the intent specific functions.