Cleaned up step intents and app.js
Removed all the conditionals in the step cases and placed them in the app specific functions.  

Created a dictionary to hold the name, index, currentIndex and previousIndex to differentiate between the different step intents 

Cleaned up app.js by moving the remainder of the code into the specific intent functions.