Updated the steps
Updated the indexing by first moving them all to nextstep.js.  Interestingly the dictionary that is set up inside of app.js and passed to nextstep.js is then updated in nextstep.js and not passed back to app.js but is updated in that file as well.  

Updated the indexing for the steps.  If a user continuously asks for the next step but is at the end of the steps the indexing will no longer increment.  Updated the indexing for requesting a specific step to not work with zero indexing because a user will not ask for step 0 for the first step.