Couple bug fixes
Fixed a bug in a edge case where the user asks for firststep then previousstep, then nextstep and got the wrong step, when it should get the second step because they're at the beginning.  

Fixed another edge case where a user asks for the 0 step and an empty object is returned.