US-80, Task-104 and Task-105
Made some minor style changes and made registration form elements as required and made some basic html validation for login and registration.

Finished post for both login and registration.  Got both working correctly with the database.More robust validation to be done next sprint if we do not completely replace it with Google sign in