US-133, Task-136 Login Session
Created an expiration of the user login to expire after 30 minutes unless a request is made in which case the session is reactivated.

Restructured the login authentication.  Checks for valid user in DB first then checks if the username and password are correct via a callback function.  Will hopefully help with sending info to the navbar to display user info.