US133, Task135 – Login and Registration Validation
Created a lot of validation for both the Login Page, the user will now be alerted (via message on response from server) if there username or password is iincorrect or does not exist.

Registration page has Express validator set up and ready to use, but I'm having issues getting it working.  For now I have auto populated messages for password and confirm-password and a more robust field requirements using html and js