US and Task – ** FUCKING FINALLY!**
Yup I F'bombed the commit title message!  

This commit is about 30 hours worth of work over the last 3 days finally paid off and every thing is working correctly! 

Session is now verified across all webpages and the buttons for Login, Registration, and Logout dynamically change as well as a Welcome Message with the users first name displayed based on if a user session exists or not.

FYI – Cookbook has always ONLY grabbed the recipes for thetoastyone on this branch.  I double checked it was nothing I did, but it was not.  

Also, buttons on the bottom of the home page do not work after changing to the handlebars template, but they're redundant with Navbar anyways.  We can definitely fix them though.  

Lastly, sorry Trevor, I added a h1 tag to the home page saying Welcome to the user before I could get the home page to dynamically update.  We can remove it later.

Going to bed, my Sprint 3 tasks are complete!