US133, Task135, US134, Task146
Made registration it's own file so that it is active when that button is selected rather than the user having to select that tab. 
TODO: Use JS or JQuery to make it active on button click.  Not sure how to do that now because it would have to dynamically change the style block and active status in addition to changing of the tabs.  So for now it's in it's own file with the same code as login.

Made it so that usernames are unique in our database by first checking if a username exists before submitting the registration form and if it exists already then provide a error message.  
TODO: repopulate registration form fields

Removed navbar.html because it is no longer used.