US133, Task136, Registration Auto Login
Made it so that after a user has successfully registered for the website they will be automatically logged in and a new user session is created with their login info.  
TODO: If we were to deploy this to the public, possibly would want the user to verify there email address first.