US-149, Task157 Welcome Message
When the browser goes below 768px the nav bar collapses to a "Hamburger" style nav bar (make browsers width smaller to see what I'm taking about).

However, when getting below 830px the Welcome Message and Login buttons messed up the NavBar becuase they were added to a table to make the Welcome message on the left and giving it a better look.  Updated the CSS to make the "Hamburger" menu appear at 830px.  More complicated than it seems because you need the rest of the website to be responsive (purpose of Bootstrap), so making the width set to a min would remove the responsiveness. 
 So a media query had to be created for the Navbar.